Fix MSVC compiler warning
reg.: 'initializing': conversion from 'size_t' to 'int', possible loss of data (C4267)

+++ src/engraving/rendering/score/systemheaderlayout.cpp
@@ -1159,7 +1159,7 @@ String SystemHeaderLayout::formattedSharedStaffLabel(staff_idx_t staffIdx, const
         }
     }
 
-    int totInstrumentCount = instrumentsMappedToFirstVoice.size() + instrumentsMappedToSecondVoice.size();
+    size_t totInstrumentCount = instrumentsMappedToFirstVoice.size() + instrumentsMappedToSecondVoice.size();
 
     SharedLabelOrientation actualOrientation;
     switch (orientation) {
